Huge improvement to API
