+++ .vscode/settings.json
@@ -46,6 +46,7 @@
     "imwrite",
     "jit",
     "Lanczos",
+    "layouted",
     "leakyrelu",
     "localstorage",
     "Luma",
